webapp/backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
numpy
